Respawn System implemented
Added to Route 7 for now.
To set a Respawn Point, make an event called "RespawnPoint" on the tile you want the player to respawn to.
To trigger the respawn system, make events that activate the script "setNewRespawnPoint" on Player Touch. Add these across an area the player will always cross, like at the start of a route.