Map Cleanups
The following redundant maps have been deleted:

Map 12, a large Dead City Shop which was not implemented. There is already a Dead City Shop (map 33) selling the same wares and is smaller

Map 35. This is an empty map that was removed, though likely created to add other map details back in. This is an empty map with no events

Map 82, a replica of the player house. This has nothing in it and redirects to Before the Forest in inaccessible areas

Map 85, another city hotel map. This is an empty small map, and there are already 4 named city hotel rooms