The Bridge and Cleave Update
Tilesets: Bridge logic should now be fully flawless. This however has meant that some areas need readjustment.

Factory Shore: Some areas on the bridge are refurbished due to the new tileset logic.

Slaughterhouse: More losetexts, and overall polishing and fixing of dialogue events. The steak dinner has also been nudged to an easier place to spot it and in a corner that feels overwise worthless to go down.